moving all highted logic to one class
